+++ pkg/storage/postgres/postgres.go
@@ -90,21 +90,13 @@ func New(uri string, cfg *sqlcommon.Config) (*Datastore, error) {
 
 // NewWithDB creates a new [Datastore] storage with the provided database connection.
 func NewWithDB(db *sql.DB, cfg *sqlcommon.Config) (*Datastore, error) {
-	if cfg.MaxOpenConns != 0 {
-		db.SetMaxOpenConns(cfg.MaxOpenConns)
-	}
-
 	if cfg.MaxIdleConns != 0 {
-		db.SetMaxIdleConns(cfg.MaxIdleConns)
+		db.SetMaxIdleConns(cfg.MaxIdleConns) // default is 2, not retaining connections(0) would be detrimental for performance
 	}
 
-	if cfg.ConnMaxIdleTime != 0 {
-		db.SetConnMaxIdleTime(cfg.ConnMaxIdleTime)
-	}
-
-	if cfg.ConnMaxLifetime != 0 {
-		db.SetConnMaxLifetime(cfg.ConnMaxLifetime)
-	}
+	db.SetMaxOpenConns(cfg.MaxOpenConns)
+	db.SetConnMaxIdleTime(cfg.ConnMaxIdleTime)
+	db.SetConnMaxLifetime(cfg.ConnMaxLifetime)
 
 	policy := backoff.NewExponentialBackOff()
 	policy.MaxElapsedTime = 1 * time.Minute